refactor: render overlay in scrollable container
