+++ libs/frontend/presentation/view/components/hoverOverlay/ClickOverlay.tsx
@@ -1,8 +1,8 @@
 import { useScroll, useScrollIntoView } from '@codelab/frontend/shared/utils'
-import React from 'react'
+import type { CSSProperties } from 'react'
+import React, { useEffect, useState } from 'react'
 import useResizeObserver from 'use-resize-observer/polyfilled'
 import type { OverlayProps } from './overlay.interface'
-import { OverlayToolbar } from './OverlayToolbar'
 
 export const ClickOverlay = ({
   content,
@@ -18,26 +18,49 @@ export const ClickOverlay = ({
   useScrollIntoView(element, renderContainer)
   useScroll()
 
-  const elementRect = element.getBoundingClientRect()
+  const [containerRect, setContainerRect] = useState(() =>
+    renderContainer.getBoundingClientRect(),
+  )
+
+  const [rect, setRect] = useState(() => element.getBoundingClientRect())
+
+  useEffect(() => {
+    setRect(element.getBoundingClientRect())
+    setContainerRect(renderContainer.getBoundingClientRect())
+  }, [element, renderContainer])
+
+  const style: CSSProperties = {
+    border: '1px solid rgb(7, 62, 78)',
+    bottom: `${rect.bottom}px`,
+    height: `${rect.height}px`,
+    left: `${rect.left - containerRect.left}px`,
+    pointerEvents: 'none',
+    position: 'fixed',
+    right: `${rect.right}px`,
+    top: `${rect.top - containerRect.top}px`,
+    width: `${rect.width}px`,
+  }
+
+  const isToolbarVisible = rect.top - containerRect.top > 0
 
   return (
-    <OverlayToolbar
-      containerProps={{
-        style: {
-          border: '1px solid rgb(7, 62, 78)',
-          maxHeight: '765px',
-          top: `${elementRect.height > 765 ? 155 : elementRect.top}px`,
-        },
-      }}
-      overlayElement={element}
-      toolbarProps={{
-        style: {
-          background: 'rgb(7, 62, 78)',
-          color: 'rgb(255, 255, 255)',
-        },
-      }}
-    >
-      {content}
-    </OverlayToolbar>
+    <div className="z-50" style={style}>
+      {content && (
+        <div
+          style={{
+            backgroundColor: 'rgb(7, 62, 78)',
+            bottom: isToolbarVisible ? '100%' : '',
+            color: 'rgb(255, 255, 255)',
+            fontSize: '0.8rem',
+            marginLeft: '-2px',
+            padding: '0.1rem 0.3rem 0.1rem 0.3rem',
+            pointerEvents: 'auto',
+            position: 'absolute',
+          }}
+        >
+          {content}
+        </div>
+      )}
+    </div>
   )
 }
