+++ libs/frontend/presentation/view/components/hoverOverlay/index.ts
@@ -1,3 +1,1 @@
 export * from './ClickOverlay'
-// export * from './HoverOverlay'
-export * from './OverlayToolbar'
